iopex-mvp-master/requirements.txt
scrapy
selenium
holmium.core
mysqlclient
scrapy-splash
beautifulsoup4
lxml
w3lib

